S-I-P-maptest/backend/requirements.txt
# Web Framework & Server
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0
python-multipart==0.0.6

# Computer Vision & AI
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# Se você usar ultralytics/YOLO, descomente:
# ultralytics==8.0.196
# torch==2.1.0
# torchvision==0.16.0
# onnx==1.15.0
# onnxruntime==1.16.3

pandas==2.3.3
scipy==1.15.3

# HTTP & Requests
aiofiles==23.2.1
httpx==0.25.2
requests==2.31.0
anyio==3.7.1
h11==0.16.0
httpcore==1.0.9
idna==3.10
sniffio==1.3.1
urllib3==2.5.0
charset-normalizer==3.4.3

# Utilities
pydantic==2.5.0
pydantic_core==2.14.1
python-dotenv==1.0.0
PyYAML==6.0.3
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
annotated-types==0.7.0
typing_extensions==4.14.1